+++ CHANGELOG.rst
@@ -2,12 +2,12 @@ Changelog
 =========
 
 
-1.3.1 (unreleased)
+2.0.0 (unreleased)
 ------------------
 
-**Bug fixes**
+**Breaking changes**
 
-- Fix support of Kinto 12
+- Requires Kinto >= 12
 
 
 1.3.0 (2018-10-15)